Going back to lib per Brian's request; keeping all doc-related scripts in separate folder;
