+++ docs/documentation/index.html
@@ -62,12 +62,12 @@
         <article><h1>json-schema-md-doc</h1>
 <h2>Generate markdown documentation for JSON Schemas</h2>
 <p>Try it out at <a href="https://brianwendt.github.io/json-schema-md-doc">brianwendt.github.io/json-schema-md-doc</a></p>
-<p>Download <a href="https://raw.githubusercontent.com/BrianWendt/json-schema-md-doc/master/src/JSONSchemaMarkdown.js">JSONSchemaMarkdown.js</a></p>
+<p>Download <a href="https://raw.githubusercontent.com/BrianWendt/json-schema-md-doc/master/docs/lib/JSONSchemaMarkdown.js">JSONSchemaMarkdown.js</a></p>
 <p><a href="https://github.com/BrianWendt/json-schema-md-doc/tree/master/samples/node">Click here</a> to see the Node example.</p>
 <p><strong>NOTE:</strong> JSONSchemaMarkdown.js supports <a href="https://json-schema.org/">json-schema.org</a> <code>draft-7</code>. Previous drafts may not generate documentation correctly.</p>
 <h2>Simple Implementation</h2>
 <p><strong>HTML</strong></p>
-<pre class="prettyprint source lang-html"><code>&lt;script src=&quot;https://brianwendt.github.io/json-schema-md-doc/src/JSONSchemaMarkdown.js&quot;>&lt;/script>
+<pre class="prettyprint source lang-html"><code>&lt;script src=&quot;https://brianwendt.github.io/json-schema-md-doc/lib/JSONSchemaMarkdown.js&quot;>&lt;/script>
 </code></pre>
 <p><strong>Javascript</strong></p>
 <pre class="prettyprint source lang-javascript"><code>// simple schema for the example
