+++ docs/index.html
@@ -4,10 +4,10 @@
         <meta charset="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
         <title>JSON Schema to Markdown</title>
-        <script src="lib/jquery-3.4.0.min.js"></script>
-        <script src="../src/JSONSchemaMarkdown.js"></script>
-        <script src="lib/index.js"></script>
-        <link href="lib/bootstrap.min.css" rel="stylesheet" />
+        <script src="doc-support/jquery-3.4.0.min.js"></script>
+        <script src="lib/JSONSchemaMarkdown.js"></script>
+        <script src="doc-support/index.js"></script>
+        <link href="doc-support/bootstrap.min.css" rel="stylesheet" />
     </head>
     <body>
         <div class="container-fluid">
@@ -35,7 +35,7 @@
                     <p><a href="./issue.html" target="_blank" class="btn btn-warning">Submit Issue</a></p>
                     <p>
                         You may fully customize your output by extending the <code>JSONSchemaMarkdown</code> class to override the rendering methods.
-                        Download <a href="../src/JSONSchemaMarkdown.js" target="_blank">JSONSchemaMarkdown.js</a>
+                        Download <a href="lib/JSONSchemaMarkdown.js" target="_blank">JSONSchemaMarkdown.js</a>
                     </p>
                     <p>Find the <a href="https://github.com/BrianWendt/json_schema_markdown" target="_blank">BrianWendt/json_schema_markdown</a> project on GitHub.</p>
                     <p class="small">License: <a href="https://github.com/BrianWendt/json_schema_markdown/blob/master/LICENSE" target="_blank">GNU General Public License v3.0</a></p>
