+++ docs/issue.html
@@ -4,9 +4,9 @@
         <meta charset="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
         <title>Submit Issue - JSON Schema to Markdown</title>
-        <script src="lib/jquery-3.4.0.min.js"></script>
-        <script src="lib/issue.js"></script>
-        <link href="lib/bootstrap.min.css" rel="stylesheet" />
+        <script src="doc-support/jquery-3.4.0.min.js"></script>
+        <script src="doc-support/issue.js"></script>
+        <link href="doc-support/bootstrap.min.css" rel="stylesheet" />
     </head>
     <body>
         <div class="container-fluid">
