+++ package.json
@@ -2,7 +2,7 @@
     "name": "json-schema-md-doc",
     "author": "Brian Wendt (https://github.com/BrianWendt)",
     "description": "Generate markdown documentation for JSON Schemas",
-    "main": "./src/JSONSchemaMarkdown.js",
+    "main": "./docs/lib/JSONSchemaMarkdown.js",
     "keywords": [
         "json-schema"
     ],
