+++ scripts/jsdoc_config.json
@@ -3,7 +3,7 @@
     "recurseDepth": 10,
     "source": {
         "include": [
-            "../src/JSONSchemaMarkdown.js"
+            "./docs/lib/JSONSchemaMarkdown.js"
         ]
     },
     "sourceType": "module",
